requirements.txt
Flask
flask_session
nltk
numpy
tensorflow
h5py
keras
matplotlib
pymongo
argon2_cffi
sklearn
